refactor(mcp-server): rename migration map to db-ui-migration-map for clarity
